dbs/weaviate/requirements.txt
weaviate-client~=3.22.0
transformers~=4.28.0
sentence-transformers~=2.2.0
pydantic~=2.0.0
pydantic-settings~=2.0.0
python-dotenv>=1.0.0
fastapi~=0.100.0
httpx>=0.24.0
aiohttp>=3.8.4
uvloop>=0.17.0
uvicorn>=0.21.0, <1.0.0
srsly>=2.4.6